refactor(cli): invert service dependency on compose files
When there's need to disable or enable some service by selectively
choosing which compose files will be part the the docker compose
command, it's easy if the dependency definition stays on the file that
define the service that's needed instead of the file that needs it.

This way, in the future we'll be able to start a node without a database
and point to an external database, for example, or start without anvil
and point to an external anvil.